[pom] Entirely rework
- Cleanup layout to match maven requirements
- Use my travis ci setup until Alex sets up travis
- Use my group id for releasing
- Fix most of what I can under Alex developer setup as tunyk.com doesn't exist.  Hint: need a real email address here.
- Add in site distribution location on Alex github

This should allow for release and will be tested as such.  First release will be semantic as 1.4.0.  If issues occur it will get raised.  Eventually it should probably just match up with underlying version of html compressor but for now need ot determine how accurate this is for release and that can be addressed at another time.