[enhance] If directory passed in doesn't exist exit gracefuly
previously this would fail at compression which wastes resources in processing if directory does not exist.  Use case where that occurs is in situations were library is used globally in super pom setup and not all modules will have anything to compress.  In such situtations, just warn it's not possible to compress and move on.  This should not kill a build when there is nothing to actually do.